+++ loss/mean_squared_error.go
@@ -6,18 +6,18 @@ import "github.com/itsmontoya/neuralnetwork/matrix"
 type MeanSquaredError struct{}
 
 // Value returns the mean squared error for predictions and targets with equal shape.
-func (m MeanSquaredError) Value(predictions, targets *matrix.Matrix) (value float64, err error) {
+func (m MeanSquaredError) Value(predictions, targets *matrix.Matrix) (value float32, err error) {
 	var (
 		rows       int
 		cols       int
-		difference float64
+		difference float32
 	)
 
 	if rows, cols, err = matrixShapePair(predictions, targets); err != nil {
 		return 0, err
 	}
 
-	err = predictions.Pairwise(targets, func(row, col int, prediction, target float64) (err error) {
+	err = predictions.Pairwise(targets, func(row, col int, prediction, target float32) (err error) {
 		difference = prediction - target
 		value += difference * difference
 		return nil
@@ -26,7 +26,7 @@ func (m MeanSquaredError) Value(predictions, targets *matrix.Matrix) (value floa
 		return 0, err
 	}
 
-	value /= float64(rows * cols)
+	value /= float32(rows * cols)
 	return value, nil
 }
 
@@ -35,7 +35,7 @@ func (m MeanSquaredError) Gradient(predictions, targets *matrix.Matrix) (gradien
 	var (
 		rows  int
 		cols  int
-		scale float64
+		scale float32
 	)
 
 	if rows, cols, err = matrixShapePair(predictions, targets); err != nil {
@@ -50,7 +50,7 @@ func (m MeanSquaredError) Gradient(predictions, targets *matrix.Matrix) (gradien
 		return nil, err
 	}
 
-	scale = 2 / float64(rows*cols)
+	scale = 2 / float32(rows*cols)
 	if err = gradient.MultiplyScalarInPlace(scale); err != nil {
 		return nil, err
 	}
