+++ matrix/elementwise_default.go
@@ -2,30 +2,30 @@
 
 package matrix
 
-func addInto(left, right, result []float64) {
+func addInto(left, right, result []float32) {
 	addIntoPure(left, right, result)
 }
 
-func addScaledInPlace(left, right []float64, scale float64) {
+func addScaledInPlace(left, right []float32, scale float32) {
 	addScaledInPlacePure(left, right, scale)
 }
 
-func subtractInto(left, right, result []float64) {
+func subtractInto(left, right, result []float32) {
 	subtractIntoPure(left, right, result)
 }
 
-func multiplyElementsInto(left, right, result []float64) {
+func multiplyElementsInto(left, right, result []float32) {
 	multiplyElementsIntoPure(left, right, result)
 }
 
-func addScalarInto(source []float64, value float64, result []float64) {
+func addScalarInto(source []float32, value float32, result []float32) {
 	addScalarIntoPure(source, value, result)
 }
 
-func multiplyScalarInto(source []float64, value float64, result []float64) {
+func multiplyScalarInto(source []float32, value float32, result []float32) {
 	multiplyScalarIntoPure(source, value, result)
 }
 
-func multiplyScalarInPlace(source []float64, value float64) {
+func multiplyScalarInPlace(source []float32, value float32) {
 	multiplyScalarInPlacePure(source, value)
 }
