+++ optimizer/constant_learning_rate.go
@@ -1,7 +1,7 @@
 package optimizer
 
 // NewConstantLearningRate constructs a schedule that always returns learningRate.
-func NewConstantLearningRate(learningRate float64) (out *ConstantLearningRate, err error) {
+func NewConstantLearningRate(learningRate float32) (out *ConstantLearningRate, err error) {
 	if err = validateLearningRate(learningRate); err != nil {
 		return nil, err
 	}
@@ -13,11 +13,11 @@ func NewConstantLearningRate(learningRate float64) (out *ConstantLearningRate, e
 
 // ConstantLearningRate returns the same learning rate for every epoch.
 type ConstantLearningRate struct {
-	learningRate float64
+	learningRate float32
 }
 
 // LearningRate returns the configured learning rate for a one-based epoch.
-func (c *ConstantLearningRate) LearningRate(epoch int) (learningRate float64, err error) {
+func (c *ConstantLearningRate) LearningRate(epoch int) (learningRate float32, err error) {
 	if err = c.validate(); err != nil {
 		return 0, err
 	}
@@ -31,7 +31,7 @@ func (c *ConstantLearningRate) LearningRate(epoch int) (learningRate float64, er
 }
 
 // Rate returns the configured learning rate.
-func (c *ConstantLearningRate) Rate() (learningRate float64) {
+func (c *ConstantLearningRate) Rate() (learningRate float32) {
 	if c == nil {
 		return 0
 	}
